moving eEmemorymap, rename files
