+++ emBODY/eBcode/arch-arm/board/amcfoc/bsp/shared/eEmemorymap.h
@@ -11,19 +11,12 @@
 #define __EEMEMORYMAP_H_
 
 
+#include "eEmemorymap_stm32h7_amcfoc.h"
 
-#define EENV_USE_MEMORYMAP_STM32H7_AMCFOC_2CM4
-
-#if defined(EENV_USE_MEMORYMAP_STM32H7_AMCFOC_2CM4)
-
-#include "eEmemorymap_stm32h7_amcfoc_2cm4.h"
-
-#endif //defined()
 
 #endif  // include-guard
 
 // - end-of-file (leave a blank line after)----------------------------------------------------------------------------
-
