+++ emBODY/eBcode/arch-arm/board/amcfoc/procs/maintainer/proj/amcfoc.2cm4-maintainer.uvprojx
@@ -1458,7 +1458,7 @@
             <TextAddressRange>0x08000000</TextAddressRange>
             <DataAddressRange>0x20000000</DataAddressRange>
             <pXoBase></pXoBase>
-            <ScatterFile>..\cfg\amcfoc.2cm4-updater.sct</ScatterFile>
+            <ScatterFile>..\cfg\amcfoc.2cm4-maintainer.sct</ScatterFile>
             <IncludeLibs></IncludeLibs>
             <IncludeLibsPath></IncludeLibsPath>
             <Misc>--diag_suppress=L6329</Misc>
